refactor: Timer
